Update Holacracy-Constitution.md
Domains are still really vague and in practice we were told how they work is that for a domain to hold for an entire organization and be able to dictate policy across all circles it either 
A. must be a policy at the GCC granting or restricting access to the domain or
B. the domain be proposed and accepted into governance at the broadest circle to the role there and then further delegated as needed.  It was expressed there needs to be some kind of chain of possession coming from the broadest organizational circle. I may be completely wrong with that interpretation but really could be interpreted either way and I'd love some clarity as to what the intent really is.